[SPARK-13331] AES support for over-the-wire encryption
Move key negotiation to AesCipher class.
