+++ common/network-common/src/main/java/org/apache/spark/network/sasl/SparkSaslClient.java
@@ -18,9 +18,7 @@
 package org.apache.spark.network.sasl;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.util.Map;
-import java.util.Properties;
 import javax.security.auth.callback.Callback;
 import javax.security.auth.callback.CallbackHandler;
 import javax.security.auth.callback.NameCallback;
@@ -34,15 +32,8 @@ import javax.security.sasl.SaslException;
 
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableMap;
-import io.netty.buffer.ByteBuf;
-import io.netty.buffer.Unpooled;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import org.apache.spark.network.client.TransportClient;
-import org.apache.spark.network.sasl.aes.AesConfigMessage;
-import org.apache.spark.network.sasl.aes.AesCipher;
-import org.apache.spark.network.util.TransportConf;
 
 import static org.apache.spark.network.sasl.SparkSaslServer.*;
 
@@ -126,38 +117,6 @@ public class SparkSaslClient implements SaslEncryptionBackend {
         saslClient = null;
       }
     }
-  }
-
-  /**
-   * Negotiate extra encryption options for SASL
-   * @param client is transport client used to connect to peer.
-   * @param conf contain client transport configuration.
-   * @throws IOException
-   * @return The object represent the result of negotiateAesSessionKey.
-   */
-  public AesCipher negotiateAesSessionKey(TransportClient client, TransportConf conf) throws IOException {
-    // Create option for negotiation
-    AesConfigMessage cipherOption = new AesConfigMessage(null, null, null, null);
-    ByteBuf buf = Unpooled.buffer(cipherOption.encodedLength());
-    cipherOption.encode(buf);
-
-    // Send option to server and decode received negotiated option
-    ByteBuffer response = client.sendRpcSync(buf.nioBuffer(), conf.saslRTTimeoutMs());
-    cipherOption = AesConfigMessage.decode(Unpooled.wrappedBuffer(response));
-
-    // Decrypt key from option. Server's outKey is client's inKey, and vice versa.
-    byte[] outKey = unwrap(cipherOption.inKey, 0, cipherOption.inKey.length);
-    byte[] inKey = unwrap(cipherOption.outKey, 0, cipherOption.outKey.length);
-
-    // Enable AES on SaslClient
-    Properties properties = new Properties();
-
-    AesCipher cipher = new AesCipher(properties, inKey, outKey,
-      cipherOption.outIv, cipherOption.inIv);
-
-    logger.debug("AES enabled for SASL client encryption.");
-
-    return cipher;
   }
 
   /**
