+++ common/network-common/src/main/java/org/apache/spark/network/sasl/SparkSaslServer.java
@@ -36,19 +36,10 @@ import javax.security.sasl.SaslServer;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 import com.google.common.collect.ImmutableMap;
-import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.handler.codec.base64.Base64;
-import org.apache.commons.crypto.cipher.CryptoCipherFactory;
-import org.apache.commons.crypto.random.CryptoRandom;
-import org.apache.commons.crypto.random.CryptoRandomFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import org.apache.spark.network.client.RpcResponseCallback;
-import org.apache.spark.network.sasl.aes.AesConfigMessage;
-import org.apache.spark.network.sasl.aes.AesCipher;
-import org.apache.spark.network.util.TransportConf;
 
 /**
  * A SASL Server for Spark which simply keeps track of the state of a single SASL session, from the
@@ -158,61 +149,6 @@ public class SparkSaslServer implements SaslEncryptionBackend {
   @Override
   public byte[] unwrap(byte[] data, int offset, int len) throws SaslException {
     return saslServer.unwrap(data, offset, len);
-  }
-
-  /**
-   * Negotiate with peer for extended options, such as using AES cipher.
-   * @param message is message receive from peer which may contains communication parameters.
-   * @param callback is rpc callback.
-   * @param conf contains transport configuration.
-   * @return Object which represent the result of negotiateAesSessionKey.
-   */
-  public AesCipher negotiateAesSessionKey(ByteBuffer message, RpcResponseCallback callback, TransportConf conf)
-   {
-    AesCipher cipher;
-
-    // Receive initial option from client
-    AesConfigMessage cipherOption = AesConfigMessage.decode(Unpooled.wrappedBuffer(message));
-    String transformation = AesCipher.TRANSFORM;
-    Properties properties = new Properties();
-
-    try {
-      // Generate key and iv
-      if (conf.saslEncryptionAesCipherKeySizeBits() % 8 != 0) {
-        throw new IllegalArgumentException("The AES cipher key size in bits should be a multiple " +
-          "of byte");
-      }
-
-      int keyLen = conf.saslEncryptionAesCipherKeySizeBits() / 8;
-      int paramLen = CryptoCipherFactory.getCryptoCipher(transformation,properties).getBlockSize();
-      byte[] inKey = new byte[keyLen];
-      byte[] outKey = new byte[keyLen];
-      byte[] inIv = new byte[paramLen];
-      byte[] outIv = new byte[paramLen];
-
-      CryptoRandom random = CryptoRandomFactory.getCryptoRandom(properties);
-      random.nextBytes(inKey);
-      random.nextBytes(outKey);
-      random.nextBytes(inIv);
-      random.nextBytes(outIv);
-
-      // Fill cipher config and send it to client.
-      cipherOption.setParameters(wrap(inKey, 0, inKey.length), inIv,
-        wrap(outKey, 0, outKey.length), outIv);
-
-      // Create cipher which will used by server
-      cipher = new AesCipher(properties, inKey, outKey, inIv, outIv);
-
-      // Send cipher option to client
-      ByteBuf buf = Unpooled.buffer(cipherOption.encodedLength());
-      cipherOption.encode(buf);
-      callback.onSuccess(buf.nioBuffer());
-    } catch (Exception e) {
-      logger.error("AES negotiation exception ", e);
-      throw Throwables.propagate(e);
-    }
-
-    return cipher;
   }
 
   /**
