+++ common/network-common/src/main/java/org/apache/spark/network/sasl/aes/AesCipher.java
@@ -26,10 +26,12 @@ import java.util.Properties;
 import javax.crypto.spec.SecretKeySpec;
 import javax.crypto.spec.IvParameterSpec;
 
+import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
-import io.netty.channel.Channel;
+import io.netty.channel.*;
+import io.netty.util.AbstractReferenceCounted;
 import org.apache.commons.crypto.cipher.CryptoCipherFactory;
 import org.apache.commons.crypto.random.CryptoRandom;
 import org.apache.commons.crypto.random.CryptoRandomFactory;
@@ -37,6 +39,8 @@ import org.apache.commons.crypto.stream.CryptoInputStream;
 import org.apache.commons.crypto.stream.CryptoOutputStream;
 import org.apache.spark.network.client.RpcResponseCallback;
 import org.apache.spark.network.sasl.SparkSaslServer;
+import org.apache.spark.network.util.ByteArrayReadableChannel;
+import org.apache.spark.network.util.ByteArrayWritableChannel;
 import org.apache.spark.network.util.TransportConf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -45,6 +49,10 @@ import org.slf4j.LoggerFactory;
  * AES cipher for encryption and decryption.
  */
 public class AesCipher {
+  private static final Logger logger = LoggerFactory.getLogger(AesCipher.class);
+  public static final String ENCRYPTION_HANDLER_NAME = "AesEncryption";
+  public static final String DECRYPTION_HANDLER_NAME = "AesDecryption";
+
   private final SecretKeySpec inKeySpec;
   private final IvParameterSpec inIvSpec;
   private final SecretKeySpec outKeySpec;
@@ -76,6 +84,11 @@ public class AesCipher {
     outIvSpec = new IvParameterSpec(outIv);
   }
 
+  public AesCipher(AesConfigMessage configMessage) throws IOException  {
+    this(new Properties(), configMessage.inKey, configMessage.outKey,
+      configMessage.inIv, configMessage.outIv);
+  }
+
   /**
    * Create AES crypto output stream
    * @param ch The underlying channel to write out.
@@ -102,6 +115,215 @@ public class AesCipher {
     }
 
     return inputStreamMap.get(ch);
+  }
+
+  /**
+   * Add handlers to channel
+   * @param ch the channel for adding handlers
+   * @throws IOException
+   */
+  public void addToChannel(Channel ch) throws IOException {
+    ch.pipeline()
+      .addFirst(ENCRYPTION_HANDLER_NAME, new AesEncryptHandler(this))
+      .addFirst(DECRYPTION_HANDLER_NAME, new AesDecryptHandler(this));
+  }
+
+  public static AesConfigMessage requestConfigMessage(TransportConf conf) {
+    int keySize = conf.saslEncryptionAesCipherKeySizeBits();
+    if (keySize % 8 != 0) {
+      throw new IllegalArgumentException("The AES cipher key size in bits should be a multiple " +
+        "of byte");
+    }
+    return new AesConfigMessage(keySize/8, null, null, null, null);
+  }
+
+  public static AesConfigMessage responseConfigMessage(AesConfigMessage configMessage){
+
+    Properties properties = new Properties();
+    int keyLen = configMessage.keySize;
+
+    try {
+      int paramLen = CryptoCipherFactory.getCryptoCipher(AesCipher.TRANSFORM, properties).getBlockSize();
+      byte[] inKey = new byte[keyLen];
+      byte[] outKey = new byte[keyLen];
+      byte[] inIv = new byte[paramLen];
+      byte[] outIv = new byte[paramLen];
+
+      CryptoRandom random = CryptoRandomFactory.getCryptoRandom(properties);
+      random.nextBytes(inKey);
+      random.nextBytes(outKey);
+      random.nextBytes(inIv);
+      random.nextBytes(outIv);
+
+      configMessage.setParameters(keyLen, inKey, inIv, outKey, outIv);
+    } catch (Exception e) {
+      logger.error("AES negotiation exception ", e);
+      throw Throwables.propagate(e);
+    }
+
+    return configMessage;
+  }
+
+  private static class AesEncryptHandler extends ChannelOutboundHandlerAdapter {
+    private ByteArrayWritableChannel byteChannel;
+    private CryptoOutputStream cos;
+
+    AesEncryptHandler(AesCipher cipher) throws IOException {
+      byteChannel = new ByteArrayWritableChannel(AesCipher.STREAM_BUFFER_SIZE);
+      cos = cipher.CreateOutputStream(byteChannel);
+    }
+
+    @Override
+    public void write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise)
+      throws Exception {
+      ctx.write(new EncryptMessage(cos, msg, byteChannel), promise);
+    }
+
+    @Override
+    public void close(ChannelHandlerContext ctx, ChannelPromise promise) throws Exception {
+      cos.close();
+      super.close(ctx, promise);
+    }
+  }
+
+  private static class AesDecryptHandler extends ChannelInboundHandlerAdapter {
+    private final Logger logger = LoggerFactory.getLogger(AesDecryptHandler.class);
+    private CryptoInputStream cis;
+    private ByteArrayReadableChannel byteChannel;
+    private long totalDecrypted;
+
+    AesDecryptHandler(AesCipher cipher) throws IOException {
+      byteChannel = new ByteArrayReadableChannel(AesCipher.STREAM_BUFFER_SIZE);
+      cis = cipher.CreateInputStream(byteChannel);
+      totalDecrypted = 0;
+    }
+
+    @Override
+    public void channelRead(ChannelHandlerContext ctx, Object data) throws Exception {
+      ByteBuf in = (ByteBuf) data;
+
+      while (in.isReadable()) {
+        byteChannel.feedData(in);
+        int i;
+        byte[] decryptedData = new byte[byteChannel.length()];
+        int offset = 0;
+        while ((i = cis.read(decryptedData, offset, decryptedData.length - offset)) > 0) {
+          offset += i;
+          if (offset >= decryptedData.length) {
+            break;
+          }
+        }
+
+        totalDecrypted += offset;
+        byteChannel.reset();
+        ctx.fireChannelRead(Unpooled.wrappedBuffer(decryptedData, 0, decryptedData.length));
+      }
+
+      in.release();
+    }
+
+    @Override
+    public void channelInactive(ChannelHandlerContext ctx) throws Exception {
+      cis.close();
+      logger.debug("{} channel decrypted {} bytes", ctx.channel(), totalDecrypted);
+      super.channelInactive(ctx);
+    }
+  }
+
+  private static class EncryptMessage extends AbstractReferenceCounted implements FileRegion {
+    private final Logger logger = LoggerFactory.getLogger(EncryptMessage.class);
+
+    private final boolean isByteBuf;
+    private final ByteBuf buf;
+    private final FileRegion region;
+    private long transferred;
+    private CryptoOutputStream cos;
+
+    // Due to streaming issue CRYPTO-125: https://issues.apache.org/jira/browse/CRYPTO-125, it has
+    // to utilize two helper ByteArrayWritableChannel for streaming. One is used to receive raw data
+    // from upper handler, another is used to store encrypted data.
+    private ByteArrayWritableChannel byteEncChannel;
+    private ByteArrayWritableChannel byteRawChannel;
+
+    private ByteBuffer currentEncrypted;
+
+    EncryptMessage(CryptoOutputStream cos, Object msg, ByteArrayWritableChannel ch) {
+      Preconditions.checkArgument(msg instanceof ByteBuf || msg instanceof FileRegion,
+        "Unrecognized message type: %s", msg.getClass().getName());
+      this.isByteBuf = msg instanceof ByteBuf;
+      this.buf = isByteBuf ? (ByteBuf) msg : null;
+      this.region = isByteBuf ? null : (FileRegion) msg;
+      this.transferred = 0;
+      byteRawChannel = new ByteArrayWritableChannel(AesCipher.STREAM_BUFFER_SIZE);
+      this.cos = cos;
+      byteEncChannel = ch;
+    }
+
+    @Override
+    public long count() {
+      return isByteBuf ? buf.readableBytes() : region.count();
+    }
+
+    @Override
+    public long position() {
+      return 0;
+    }
+
+    @Override
+    public long transfered() {
+      return transferred;
+    }
+
+    @Override
+    public long transferTo(WritableByteChannel target, long position) throws IOException {
+      Preconditions.checkArgument(position == transfered(), "Invalid position.");
+
+      do {
+        if(currentEncrypted == null) {
+          encryptMore();
+        }
+        int byteWritten = currentEncrypted.remaining();
+        target.write(currentEncrypted);
+        byteWritten -= currentEncrypted.remaining();
+        transferred += byteWritten;
+        if(!currentEncrypted.hasRemaining()) {
+          currentEncrypted = null;
+          byteEncChannel.reset();
+        }
+      } while (transferred < count());
+
+      logger.debug("{} transferred total {} bytes", target, transferred);
+
+      return transferred;
+    }
+
+    public void encryptMore() throws IOException {
+
+      byteRawChannel.reset();
+      if (isByteBuf) {
+        int copied = byteRawChannel.write(buf.nioBuffer());
+        buf.skipBytes(copied);
+      } else {
+        region.transferTo(byteRawChannel, region.transfered());
+      }
+      cos.write(byteRawChannel.getData(), 0, byteRawChannel.length());
+      cos.flush();
+
+      currentEncrypted = ByteBuffer.wrap(byteEncChannel.getData(),
+        0, byteEncChannel.length());
+    }
+
+    @Override
+    protected void deallocate() {
+      byteRawChannel.reset();
+      byteEncChannel.reset();
+      if (region != null) {
+        region.release();
+      }
+      if (buf != null) {
+        buf.release();
+      }
+    }
   }
 
 }
